removed unnecessary floating point bloat

+++ source/ADC.h
@@ -3,7 +3,7 @@
 
 void initADC(void);
 uint16_t getADC(uint8_t chan);
-float voltage(uint8_t chan, uint8_t hi);
+uint16_t voltage(uint8_t chan, uint8_t hi);
